fixed file paths, nested functionality
